+++ core/util/src/main/kotlin/com/teamyg/parfait/core/util/extensions/Context.kt
@@ -1,11 +1,13 @@
-package com.teamyg.parfait.feature.gallery.impl.utils.extension
+package com.teamyg.parfait.core.util.extensions
 
 import android.content.Context
 import android.content.Intent
+import android.content.pm.PackageManager
 import android.net.Uri
 import android.provider.Settings
+import androidx.core.content.ContextCompat
 
-internal fun Context.buildAppSettingsIntent(): Intent {
+fun Context.buildAppSettingsIntent(): Intent {
     val packageName = this.packageName
 
     return Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS).apply {
@@ -13,3 +15,6 @@ internal fun Context.buildAppSettingsIntent(): Intent {
         addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
     }
 }
+
+fun Context.isGrantedPermission(permission: String): Boolean =
+    ContextCompat.checkSelfPermission(this, permission) == PackageManager.PERMISSION_GRANTED
