+++ feature/camera/impl/src/main/java/com/teamyg/parfait/feature/camera/impl/route/CustomCameraRoute.kt
@@ -1,7 +1,8 @@
 package com.teamyg.parfait.feature.camera.impl.route
 
 import android.Manifest
-import android.content.pm.PackageManager
+import android.app.Activity
+import android.content.Context
 import androidx.activity.compose.LocalActivity
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
@@ -24,11 +25,13 @@ import androidx.navigation3.runtime.result.LocalResultEventBus
 import com.teamyg.parfait.feature.camera.impl.component.CameraPreviewComponent
 import com.teamyg.parfait.feature.camera.impl.screen.CustomCameraScreen
 import com.teamyg.parfait.feature.camera.impl.util.CameraFileProvider
-import com.teamyg.parfait.feature.camera.impl.util.CameraPermissionUtil
 import com.teamyg.parfait.feature.camera.impl.viewmodel.CustomCameraEffect
 import com.teamyg.parfait.feature.camera.impl.viewmodel.CustomCameraIntent
 import com.teamyg.parfait.feature.camera.impl.viewmodel.CustomCameraViewModel
 import com.teamyg.parfait.core.navigation.Navigator
+import com.teamyg.parfait.core.util.extensions.buildAppSettingsIntent
+import com.teamyg.parfait.core.util.extensions.isGrantedPermission
+import com.teamyg.parfait.core.util.extensions.shouldShowRationale
 
 @Composable
 internal fun CustomCameraRoute(
@@ -36,8 +39,8 @@ internal fun CustomCameraRoute(
     modifier: Modifier = Modifier,
     viewModel: CustomCameraViewModel = hiltViewModel(),
 ) {
-    val context = LocalContext.current
-    val activity = LocalActivity.current
+    val activity: Activity? = LocalActivity.current
+    val context: Context = activity ?: LocalContext.current
 
     val resultEventBus = LocalResultEventBus.current
     val state by viewModel.state.collectAsStateWithLifecycle()
@@ -50,10 +53,7 @@ internal fun CustomCameraRoute(
         viewModel.processIntent(
             CustomCameraIntent.OnPermissionRequestResult(
                 granted = granted,
-                shouldShowRationale = CameraPermissionUtil.shouldShowRationale(
-                    activity = activity,
-                    permission = Manifest.permission.CAMERA,
-                ),
+                shouldShowRationale = activity?.shouldShowRationale(Manifest.permission.CAMERA) == true,
             ),
         )
     }
@@ -66,7 +66,7 @@ internal fun CustomCameraRoute(
                 }
 
                 is CustomCameraEffect.OpenAppSettings -> {
-                    CameraPermissionUtil.openAppSettings(context)
+                    context.startActivity(context.buildAppSettingsIntent())
                 }
 
                 is CustomCameraEffect.CaptureImage -> {
@@ -104,10 +104,7 @@ internal fun CustomCameraRoute(
     }
 
     LifecycleResumeEffect(Unit) {
-        val granted = ContextCompat.checkSelfPermission(
-            context,
-            Manifest.permission.CAMERA,
-        ) == PackageManager.PERMISSION_GRANTED
+        val granted = context.isGrantedPermission(permission = Manifest.permission.CAMERA)
 
         viewModel.processIntent(CustomCameraIntent.OnPermissionResult(granted))
         onPauseOrDispose { }
